feat: update database models and enhance admin conversation handling
- Changed the type of `telegram_id` in the Customer model from Integer to BigInteger to accommodate larger user IDs.
- Added a new method `_reset_conversation` in AdminHandler to reset user conversation state and clear user data, improving user experience.
- Updated command handlers to utilize the new conversation reset functionality, ensuring a consistent approach to managing user interactions.
- Simplified error handling in the error handler by replacing dynamic text with a static button label for better clarity.

+++ src/db/models.py
@@ -8,7 +8,7 @@ Properly defined models with correct Base class and type annotations.
 from datetime import datetime
 from typing import Any, List, Optional, Type
 
-from sqlalchemy import JSON, Boolean, DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import JSON, Boolean, DateTime, Float, ForeignKey, Integer, BigInteger, String, Text
 from sqlalchemy.orm import (
     DeclarativeMeta,
     Mapped,
@@ -32,7 +32,7 @@ class Customer(Base):
     __tablename__ = "customers"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    telegram_id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False)
+    telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False)
     name: Mapped[str] = mapped_column(String(100), nullable=False)
     phone: Mapped[Optional[str]] = mapped_column(String(20), nullable=True)
     language: Mapped[Optional[str]] = mapped_column(String(10), nullable=True)
